Move SocialLogin to the Smr namespace
* Use the PSR-4 autoloader instead of our custom one.
* Rename the namespace SocialLogins -> SocialLogin
* Move the SocialLogin class into the SocialLogin namespace.
